requirements/development.txt
awscli==1.29.76
aiohttp==3.9.1
aiofiles==23.2.1
boto3==1.28.76
celery==5.3.1
Collectfast==2.2.0
dj-database-url==1.2.0
Django==4.2.7
django-autoslug==1.9.8
django-celery-results==2.5.1
django-compressor==4.3.1
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-dotenv==1.4.2
django-elasticsearch-dsl==7.3
django-enumerify==0.5.2
django-extensions==3.2.3
django-filter==23.2
django-heartbeat==2.1.0
django-redis==5.3.0
#django-redis-cache==1.6.4
django-redis-sessions==0.6.2
django-reversion==5.0.4
django-silk==5.0.4
django-simple-history==3.4.0
django-storages==1.13.2
django-test-without-migrations==0.6
django-user-agents==0.4.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
# djangorestframework-simplejwt @ git+https://github.com/jazzband/djangorestframework-simplejwt.git@master
dotmap==1.3.30
drf-spectacular==0.25.1
elastic-apm==6.13.2
elasticsearch==7.0.0
elasticsearch-dsl==7.4.0
exponent-server-sdk==2.0.0
factory-boy==3.3.0
Faker==19.6.2
flower==2.0.0
future==0.18.2
isoweek==1.3.3
icecream==2.1.3
jellyfish==0.9.0
locust==2.15.1
mattermost==6.5.0
nest-asyncio==1.5.6
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
pdfkit==1.0.0
Pillow==9.5.0
pprintpp==0.4.0
psycopg2-binary==2.9.5
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyPDF2==3.0.1
python-barcode==0.15.1
python-dateutil==2.8.2
python-slugify==7.0.0
pymsteams==0.2.2
pytz==2022.7
redis==4.4.0
requests==2.28.1
selenium==4.7.2
six==1.16.0
slackclient==2.9.4
sorl-thumbnail==12.9.0
tabulate==0.9.0
tblib==1.7.0
tqdm==4.64.1
urllib3==1.26.13
validator-collection==1.5.0
WeasyPrint==60.2
Werkzeug==2.2.2
xlrd==2.0.1
yaspin==2.2.0
onesignal-sdk==2.0.0
django-json-field-schema-validator==0.0.2